Fix things brought up by PHPStan : the main change is the split up of some traits to have a better coherence in what methods can be called based on the context they are called
