+++ src/Command/Create.php
@@ -23,6 +23,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 class Create extends CreateCommand
 {
 
+    use CommandTrait;
     use ConfigurationTrait {
         execute as parentExecute;
     }
