+++ src/Command/Dump.php
@@ -13,7 +13,7 @@ namespace Migrations\Command;
 
 use Cake\Datasource\ConnectionManager;
 use Migrations\ConfigurationTrait;
-use Migrations\Shell\Task\SnapshotTrait;
+use Migrations\TableFinderTrait;
 use Phinx\Console\Command\AbstractCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -27,8 +27,16 @@ use Symfony\Component\Console\Output\OutputInterface;
 class Dump extends AbstractCommand
 {
 
+    use CommandTrait;
     use ConfigurationTrait;
-    use SnapshotTrait;
+    use TableFinderTrait;
+
+    /**
+     * Output object.
+     *
+     * @var \Symfony\Component\Console\Output\OutputInterface
+     */
+    protected $output;
 
     /**
      * {@inheritdoc}
@@ -41,8 +49,8 @@ class Dump extends AbstractCommand
                 '%sDumps the current schema of the database to be used while baking a diff%s',
                 PHP_EOL,
                 PHP_EOL
-            ));
-        $this->addOption('plugin', 'p', InputOption::VALUE_REQUIRED, 'The plugin the file should be created for')
+            ))
+            ->addOption('plugin', 'p', InputOption::VALUE_REQUIRED, 'The plugin the file should be created for')
             ->addOption('connection', 'c', InputOption::VALUE_REQUIRED, 'The datasource connection to use')
             ->addOption('source', 's', InputOption::VALUE_REQUIRED, 'The folder where migrations are in');
     }
