+++ src/Command/Rollback.php
@@ -21,6 +21,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 class Rollback extends RollbackCommand
 {
 
+    use CommandTrait;
     use ConfigurationTrait {
         execute as parentExecute;
     }
