+++ src/ConfigurationTrait.php
@@ -134,8 +134,8 @@ trait ConfigurationTrait
      * that was configured for the configuration.
      *
      * @param string $driver The driver name as configured for the CakePHP app.
-     * @return \Phinx\Config\Config
-     * @throws \InvalidArgumentexception when it was not possible to infer the information
+     * @return string Name of the adapter.
+     * @throws \InvalidArgumentException when it was not possible to infer the information
      * out of the provided database configuration
      */
     public function getAdapterName($driver)
@@ -155,77 +155,7 @@ trait ConfigurationTrait
                 return 'sqlsrv';
         }
 
-        throw new \InvalidArgumentexception('Could not infer database type from driver');
-    }
-
-    /**
-     * Overrides the action execute method in order to vanish the idea of environments
-     * from phinx. CakePHP does not believe in the idea of having in-app environments
-     *
-     * @param \Symfony\Component\Console\Input\InputInterface $input the input object
-     * @param \Symfony\Component\Console\Output\OutputInterface $output the output object
-     * @return void
-     */
-    protected function execute(InputInterface $input, OutputInterface $output)
-    {
-        $this->beforeExecute($input, $output);
-        parent::execute($input, $output);
-    }
-
-    /**
-     * Overrides the action execute method in order to vanish the idea of environments
-     * from phinx. CakePHP does not believe in the idea of having in-app environments
-     *
-     * @param \Symfony\Component\Console\Input\InputInterface $input the input object
-     * @param \Symfony\Component\Console\Output\OutputInterface $output the output object
-     * @return void
-     */
-    protected function beforeExecute(InputInterface $input, OutputInterface $output)
-    {
-        $this->setInput($input);
-        $this->addOption('--environment', '-e', InputArgument::OPTIONAL);
-        $input->setOption('environment', 'default');
-    }
-
-    /**
-     * Sets the input object that should be used for the command class. This object
-     * is used to inspect the extra options that are needed for CakePHP apps.
-     *
-     * @param \Symfony\Component\Console\Input\InputInterface $input the input object
-     * @return void
-     */
-    public function setInput(InputInterface $input)
-    {
-        $this->input = $input;
-    }
-
-    /**
-     * A callback method that is used to inject the PDO object created from phinx into
-     * the CakePHP connection. This is needed in case the user decides to use tables
-     * from the ORM and executes queries.
-     *
-     * @param \Symfony\Component\Console\Input\InputInterface $input the input object
-     * @param \Symfony\Component\Console\Output\OutputInterface $output the output object
-     * @return void
-     */
-    public function bootstrap(InputInterface $input, OutputInterface $output)
-    {
-        parent::bootstrap($input, $output);
-        $name = $this->getConnectionName($input);
-        $this->connection = $name;
-        ConnectionManager::alias($name, 'default');
-        $connection = ConnectionManager::get($name);
-
-        $manager = $this->getManager();
-
-        if (!$manager instanceof CakeManager) {
-            $this->setManager(new CakeManager($this->getConfig(), $input, $output));
-        }
-        $env = $this->getManager()->getEnvironment('default');
-        $adapter = $env->getAdapter();
-        if (!$adapter instanceof CakeAdapter) {
-            $env->setAdapter(new CakeAdapter($adapter, $connection));
-        }
+        throw new \InvalidArgumentException('Could not infer database type from driver');
     }
 
     /**
