+++ src/Shell/Task/MigrationSnapshotTask.php
@@ -17,6 +17,7 @@ use Cake\Core\Configure;
 use Cake\Datasource\ConnectionManager;
 use Cake\Event\Event;
 use Cake\Event\EventManager;
+use Migrations\TableFinderTrait;
 use Migrations\Util\UtilTrait;
 
 /**
@@ -24,8 +25,10 @@ use Migrations\Util\UtilTrait;
  */
 class MigrationSnapshotTask extends SimpleMigrationTask
 {
-    use UtilTrait;
+
     use SnapshotTrait;
+    use TableFinderTrait;
+    use UtilTrait;
 
     /**
      * {@inheritDoc}
